pyproject.toml
[project]
name = "oc-p2-fashion-trend-intelligence"
version = "0.1.0"
description = ""
authors = [
    {name = "Xavier COULON",email = "xcoulon.mail@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jupyterlab (>=4.4.6,<5.0.0)",
    "ipykernel (>=6.30.1,<7.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "matplotlib (>=3.10.5,<4.0.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "huggingface-hub (>=0.34.4,<0.35.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "ipywidgets (>=8.1.7,<9.0.0)"
]

[tool.poetry]
packages = [{include = "oc_p2_fashion_trend_intelligence", from = "src"}]


[tool.poetry.group.dev.dependencies]
nbstripout = "^0.8.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
